pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "psycopg2-binary>=2.9.11",
    "python-telegram-bot>=22.7",
]
